Refactor authentication error handling; move error handling logic to ServiceHelper and reorganize type imports
